ml/requirements.txt
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0

